requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
colorlog==6.10.1
defusedxml==0.7.1
dill==0.4.1
docling==2.91.0
docling-core==2.74.1
docling-ibm-models==3.13.2
docling-parse==5.10.1
et_xmlfile==2.0.0
Faker==40.15.0
fastapi==0.136.0
python-multipart==0.0.20
filelock==3.29.0
filetype==1.2.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.12
Jinja2==3.1.6
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
latex2mathml==3.81.0
lxml==6.1.0
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.4
omegaconf==2.3.0
opencv-python==4.13.0.92
openpyxl==3.1.5
packaging==26.2
pandas==3.0.2
pillow==12.2.0
pluggy==1.6.0
polyfactory==3.3.0
psutil==7.2.2
pyclipper==1.4.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
pylatexenc==2.10
pypdfium2==5.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-pptx==1.0.2
pywin32==311
PyYAML==6.0.3
rapidocr==3.8.1
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
rtree==1.4.1
safetensors==0.7.0
scipy==1.17.1
semchunk==3.2.5
setuptools==81.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
sympy==1.14.0
tabulate==0.10.0
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.6.2
tree-sitter==0.25.2
tree-sitter-c==0.24.2
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
websockets==16.0
xlsxwriter==3.2.9
